+++ lib/kettle/soup/cover/formatters.rb
@@ -6,9 +6,7 @@ module Kettle
       module_function
 
       def configure_formatters!
-        if Constants::TURBO_TESTS_WORKER
-          SimpleCov.at_exit { SimpleCov.result }
-        elsif Constants::MULTI_FORMATTERS
+        if Constants::MULTI_FORMATTERS
           Kettle::Soup::Cover::Loaders.load_formatters
         else
           require "simplecov-html"