+++ spec/kettle/soup/cover/rakelib/turbo_tests_spec.rb
@@ -52,22 +52,22 @@ RSpec.describe "rake turbo_tests:setup" do
     load File.join(gem_root, "lib/kettle/soup/cover/rakelib/turbo_tests.rake")
   end
 
-  it "clears turbo_tests2 coverage when coverage is enabled" do
+  it "clears shared coverage output when coverage is enabled" do
     allow(Kettle::Soup::Cover).to receive(:turbo_tests_coverage?).and_return(true)
-    allow(Kettle::Soup::Cover).to receive(:clear_turbo_tests_coverage_dir!)
+    allow(Kettle::Soup::Cover).to receive(:clear_coverage_dir!)
 
     rake["turbo_tests:setup"].invoke
 
-    expect(Kettle::Soup::Cover).to have_received(:clear_turbo_tests_coverage_dir!)
+    expect(Kettle::Soup::Cover).to have_received(:clear_coverage_dir!)
   end
 
-  it "does not clear turbo_tests2 coverage when coverage is disabled" do
+  it "does not clear shared coverage output when coverage is disabled" do
     allow(Kettle::Soup::Cover).to receive(:turbo_tests_coverage?).and_return(false)
-    allow(Kettle::Soup::Cover).to receive(:clear_turbo_tests_coverage_dir!)
+    allow(Kettle::Soup::Cover).to receive(:clear_coverage_dir!)
 
     rake["turbo_tests:setup"].invoke
 
-    expect(Kettle::Soup::Cover).not_to have_received(:clear_turbo_tests_coverage_dir!)
+    expect(Kettle::Soup::Cover).not_to have_received(:clear_coverage_dir!)
   end
 end
 
@@ -81,22 +81,22 @@ RSpec.describe "rake turbo_tests2:setup" do
     load File.join(gem_root, "lib/kettle/soup/cover/rakelib/turbo_tests.rake")
   end
 
-  it "clears turbo_tests2 coverage when coverage is enabled" do
+  it "clears shared coverage output when coverage is enabled" do
     allow(Kettle::Soup::Cover).to receive(:turbo_tests_coverage?).and_return(true)
-    allow(Kettle::Soup::Cover).to receive(:clear_turbo_tests_coverage_dir!)
+    allow(Kettle::Soup::Cover).to receive(:clear_coverage_dir!)
 
     rake["turbo_tests2:setup"].invoke
 
-    expect(Kettle::Soup::Cover).to have_received(:clear_turbo_tests_coverage_dir!)
+    expect(Kettle::Soup::Cover).to have_received(:clear_coverage_dir!)
   end
 
-  it "does not clear turbo_tests2 coverage when coverage is disabled" do
+  it "does not clear shared coverage output when coverage is disabled" do
     allow(Kettle::Soup::Cover).to receive(:turbo_tests_coverage?).and_return(false)
-    allow(Kettle::Soup::Cover).to receive(:clear_turbo_tests_coverage_dir!)
+    allow(Kettle::Soup::Cover).to receive(:clear_coverage_dir!)
 
     rake["turbo_tests2:setup"].invoke
 
-    expect(Kettle::Soup::Cover).not_to have_received(:clear_turbo_tests_coverage_dir!)
+    expect(Kettle::Soup::Cover).not_to have_received(:clear_coverage_dir!)
   end
 end
 # rubocop:enable RSpec/MultipleDescribes